refactor: dashboard components and restructure file organization
- Deleted unused components: WalletIcon, WalletList, WalletListCard, WalletSummary
- Moved wallet-related components to a new 'wallet' directory for better organization
- Created a new layout component: DashboardLayout to encapsulate NavBar and SideBar
- Updated Dashboard and TransactionManagement pages to use the new layout structure
- Implemented a new SideBar and NavBar with improved functionality and user experience
- Added navigation constants for better management of navigation items
- Enhanced WalletList and WalletListCard components with improved styling and functionality
- Updated StatsOverview to provide a comprehensive financial overview
