+++ frontend/src/features/dashboard/components/layout/DashboardLayout.jsx
@@ -1,6 +1,7 @@
 import PropTypes from 'prop-types';
-import { NavBar, SideBar } from '../index.js';
-import { UserService } from '../../../shared';
+import NavBar from './NavBar.jsx';
+import SideBar from './SideBar.jsx';
+import { UserService } from '../../../../shared';
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 
@@ -14,7 +15,7 @@ export default function DashboardLayout({ children }) {
             <input id="drawer" type="checkbox" className="drawer-toggle" />
             <div className="drawer-content flex flex-col items-start justify-start h-full">
                 {/* Navbar */}
-                {NavBar(navigate, username)}
+                <NavBar navigate={navigate} username={username} />
                 
                 {/* Main content */}
                 <main className="flex-1 w-full">
@@ -25,7 +26,7 @@ export default function DashboardLayout({ children }) {
             {/* Sidebar */}
             <div className="drawer-side">
                 <label htmlFor="drawer" aria-label="close sidebar" className="drawer-overlay"></label>
-                {SideBar()}
+                <SideBar />
             </div>
         </div>
     );
