+++ frontend/src/features/dashboard/components/layout/SideBar.jsx
@@ -1,61 +1,16 @@
-import { HouseIcon, ReceiptIcon, WaveSawtoothIcon, ListMagnifyingGlassIcon, GearIcon, QuestionIcon, WalletIcon } from "@phosphor-icons/react"
 import { useState, useEffect } from "react"
 import { useNavigate, useLocation } from "react-router-dom"
 import SideBarItem from "./SideBarItem"
+import { NAVIGATION_ITEMS, DEFAULT_ACTIVE_ITEM } from "../../constants/navigation.jsx"
 
 export default function SideBar() {
-    const [activeItem, setActiveItem] = useState("總覽")
+    const [activeItem, setActiveItem] = useState(DEFAULT_ACTIVE_ITEM)
     const navigate = useNavigate()
     const location = useLocation()
 
-    const menuItems = [
-        {
-            icon: <HouseIcon size={24} />,
-            text: "總覽",
-            id: "overview",
-            path: "/dashboard"
-        },
-        {
-            icon: <WalletIcon size={24} />,
-            text: "錢包管理",
-            id: "wallets",
-            path: "/wallets"
-        },
-        {
-            icon: <ReceiptIcon size={24} />,
-            text: "交易管理",
-            id: "transactions",
-            path: "/transactions"
-        },
-        {
-            icon: <WaveSawtoothIcon size={24} />,
-            text: "趨勢圖",
-            id: "trends",
-            path: "/trends"
-        },
-        {
-            icon: <ListMagnifyingGlassIcon size={24} />,
-            text: "所有紀錄",
-            id: "records",
-            path: "/records"
-        },
-        {
-            icon: <GearIcon size={24} />,
-            text: "設定",
-            id: "settings",
-            path: "/settings"
-        },
-        {
-            icon: <QuestionIcon size={24} />,
-            text: "幫助",
-            id: "help",
-            path: "/help"
-        }
-    ]
-
     useEffect(() => {
         const currentPath = location.pathname
-        const matchedItem = menuItems.find(item => item.path === currentPath)
+        const matchedItem = NAVIGATION_ITEMS.find(item => item.path === currentPath)
         if (matchedItem) {
             setActiveItem(matchedItem.text)
         }
@@ -78,7 +33,7 @@ export default function SideBar() {
             </div>
             {/* Sidebar content here */}
             <ul className='menu p-4 w-full bg-base-100 text-base-content gap-2'>
-                {menuItems.map((item) => (
+                {NAVIGATION_ITEMS.map((item) => (
                     <SideBarItem
                         key={item.id}
                         icon={item.icon}
