+++ frontend/src/features/dashboard/components/wallet/WalletIcon.jsx
